backend/requirements.txt
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
google-ads==28.4.1
google-api-core==2.28.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.76.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
oauthlib==3.3.1
proto-plus==1.26.1
protobuf==6.33.2
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
SQLAlchemy==2.0.45
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.4
